refactor(minecraft): delete ModMinecraft, move remaining method to ModDownload
