+++ Plain Craft Launcher 2/Modules/Minecraft/ModDownload.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Net;
 using System.Text;
+using System.Text.Json.Nodes;
 using PCL.Core.App;
 using PCL.Core.App.Localization;
 using PCL.Core.IO.Net.Http;
@@ -388,7 +389,7 @@ public static class ModDownload
                                       !_DlClientListMojangMain_IsHinted)
             {
                 _DlClientListMojangMain_IsHinted = true;
-                ModMinecraft.McDownloadClientUpdateHint(version, json);
+                McDownloadClientUpdateHint(version, json);
             }
 
             States.Tool.LastSnapshot = version ?? "Nothing";
@@ -400,7 +401,7 @@ public static class ModDownload
                                       !_DlClientListMojangMain_IsHinted)
             {
                 _DlClientListMojangMain_IsHinted = true;
-                ModMinecraft.McDownloadClientUpdateHint(version, json);
+                McDownloadClientUpdateHint(version, json);
             }
 
             States.Tool.LastRelease = version;
@@ -2554,4 +2555,48 @@ public static class ModDownload
         new("Legacy Fabric API List Loader", task => task.output = ModComp.CompFilesGet("legacy-fabric-api", false));
 
     #endregion
+
+    /// <summary>
+    ///     发送 Minecraft 更新提示。
+    /// </summary>
+    public static void McDownloadClientUpdateHint(string versionName, JsonObject json)
+    {
+        try
+        {
+            // 获取对应版本
+            JsonNode version = null;
+            foreach (var Token in json["versions"].AsArray())
+                if (Token["id"] is not null && (Token["id"].ToString() ?? "") == (versionName ?? ""))
+                {
+                    version = Token;
+                    break;
+                }
+
+            // 进行提示
+            if (version is null)
+                return;
+            var time = version["releaseTime"].ToObject<DateTime>();
+            var msgBoxText = Lang.Text("Minecraft.Update.NewVersion", versionName) + "\r\n" +
+                             ((DateTime.Now - time).TotalDays > 1d
+                                 ? Lang.Text("Minecraft.Update.UpdateTime") + Lang.Date(time)
+                                 : Lang.Text("Minecraft.Update.UpdatedAt") + Lang.TimeSpan(time - DateTime.Now));
+            var msgResult = ModMain.MyMsgBox(msgBoxText, Lang.Text("Minecraft.Update.Title"),
+                Lang.Text("Common.Action.Confirm"), Lang.Text("Common.Action.Download"),
+                (DateTime.Now - time).TotalHours > 3d ? Lang.Text("Common.Action.UpdateLog") : "",
+                button3Action: () => ModDownloadLib.McUpdateLogShow(version));
+            // 弹窗结果
+            if (msgResult == 2)
+                // 下载
+                ModBase.RunInUi(() =>
+                {
+                    PageDownloadInstall.mcVersionWaitingForSelect = versionName;
+                    ModMain.frmMain.PageChange(FormMain.PageType.Download, FormMain.PageSubType.DownloadInstall);
+                });
+        }
+
+        catch (Exception ex)
+        {
+            ModBase.Log(ex, Lang.Text("Minecraft.Error.UpdateNotify", versionName ?? "Nothing"), ModBase.LogLevel.Feedback);
+        }
+    }
 }
